feat(autoware_tensorrt_vad): rename net.h into net.hpp
