+++ AV-Solutions/vad-trt/app/autoware_tensorrt_vad/include/autoware/tensorrt_vad/vad_model.hpp
@@ -24,7 +24,7 @@
 #include <cuda_runtime.h>
 #include <NvInfer.h>
 #include <dlfcn.h>
-#include "net.h"
+#include "net.hpp"
 #include <map>
 
 #include <autoware/tensorrt_common/tensorrt_common.hpp>
